Implemented a general signin controller
