+++ src/controllers/client.controller.js
@@ -8,36 +8,15 @@ module.exports = {
       const { body } = req;
       const client = await Client.create(body);
 
-      const token = jwt.sign({ userId: client._id }, process.env.SECRET, {
-        expiresIn: 60 * 60 * 24
-      });
+      const token = jwt.sign(
+        { userId: client._id, userType: 'client' },
+        process.env.SECRET, 
+        { expiresIn: 60 * 60 * 24 }
+      );
 
       res.status(201).json({ message: 'client created successfully', token });
     } catch (error) {
       res.status(400).json({ message: error });
-    }
-  },
-  async signin(req, res) {
-    try {
-      const { email, password } = req.body;
-      const client = await Client.findOne({ email });
-
-      if(!client) {
-        throw Error('Usuario o contraseña inválida');
-      };
-
-      const isValid = await bcrypt.compare(password, client.password);
-
-      if(!isValid) {
-        throw Error('Usuario o contraseña inválida');
-      };
-
-      const token = jwt.sign({ userId: user._id }, process.env.SECRET, {
-        expiresIn: 60 * 60 * 24,
-      });
-      res.status(201).json({token});
-    } catch(error) {
-      res.status(401).json({ message: error.message })
     }
   },
   async list(req, res) {
